Initial implementation of async dependencies
Based on tracking certain (not all: requires specific launch call on
the user's part) asyncs using a monotonically increasing sequence
number. A lazy approach is used for enqueueing tasks for execution when
they have a dependency, wherein they are staged locally on the origin
side until a completion notification is received from the rank where
their dependency has executed (the notification contains the sequence
number of the dependency). Once the latter is received, the dependent
task will be moved to the target as ususal.

Need a sane way to garbage collect sequence numbers that are no longer
used. For example, if a particular task has already completed and its
associated notification has been received by the origin, the user may
still be holding onto the sequence number "handle" and be launching
dependent tasks.  Thus, it is not safe to remove the sequence number
from the set of known completed tasks until we are sure that the user
will never use it again (i.e. the handle goes out of scope and all
dependents have executed).

Sounds like we need some kind of reference counting, both to track
whether the handle is still in scope anywhere and whether there remain
any outstanding dependent tasks.