Separate task representation for messages
The AM-like message used to convey task info onto the target is now
represented with a different structure from what is used for outgoing
enqueued tasks. This is good, because there is no need to tx info on
dependencies to the target (as they are in general already resolved due
to the lazy communication approach used).

The latter enables us to do some more exciting things with dependency
resolution that we do not want to bleed over (superfluously) into our
task messages (for example, multi-dependencies for a DAG-like model).